Implemented supervised and unsupervised clients with unit tests
