+++ unit/client/test_supervised_client.py
@@ -9,7 +9,7 @@ from unit.utils.async_test import AsyncTestCase
 from dp_fasttext.client.testing.mock_client import MockClient, mock_sentence_vector, mock_labels_api, mock_invalid_response
 
 
-class ClientTestCase(TestCase, AsyncTestCase):
+class SupervisedClientTestCase(TestCase, AsyncTestCase):
 
     def test_get_sentence_vector(self):
         """
@@ -35,14 +35,14 @@ class ClientTestCase(TestCase, AsyncTestCase):
             # Init mock client
             async with MockClient() as client:
                 # Mock out _post
-                client._post = MagicMock(return_value=return_fn())
+                client.post = MagicMock(return_value=return_fn())
 
                 expected_uri = "/supervised/sentence/vector"
 
                 # Make the call
-                vector = await client.get_sentence_vector(query, headers=headers)
+                vector = await client.supervised.get_sentence_vector(query, headers=headers)
 
-                client._post.assert_called_with(expected_uri, data, headers=headers)
+                client.post.assert_called_with(expected_uri, data, headers=headers)
 
         self.run_async(async_test_function)
 
@@ -70,15 +70,15 @@ class ClientTestCase(TestCase, AsyncTestCase):
             # Init mock client
             async with MockClient() as client:
                 # Mock out _post
-                client._post = MagicMock(return_value=return_fn())
+                client.post = MagicMock(return_value=return_fn())
 
                 expected_uri = "/supervised/sentence/vector"
 
                 # Make the call and assert exception raised
                 with self.assertRaises(Exception) as context:
-                    vector = await client.get_sentence_vector(query, headers=headers)
+                    vector = await client.supervised.get_sentence_vector(query, headers=headers)
                     self.assertIn("Invalid response for method 'get_sentence_vector'", str(context))
-                client._post.assert_called_with(expected_uri, data, headers=headers)
+                client.post.assert_called_with(expected_uri, data, headers=headers)
 
         self.run_async(async_test_function)
 
@@ -111,14 +111,14 @@ class ClientTestCase(TestCase, AsyncTestCase):
             # Init mock client
             async with MockClient() as client:
                 # Mock out _post
-                client._post = MagicMock(return_value=return_fn())
+                client.post = MagicMock(return_value=return_fn())
 
                 expected_uri = "/supervised/predict"
 
                 # Make the call
-                labels, probabilities = await client.predict(query, num_labels, threshold, headers=headers)
+                labels, probabilities = await client.supervised.predict(query, num_labels, threshold, headers=headers)
 
-                client._post.assert_called_with(expected_uri, data, headers=headers)
+                client.post.assert_called_with(expected_uri, data, headers=headers)
 
         self.run_async(async_test_function)
 
@@ -151,14 +151,14 @@ class ClientTestCase(TestCase, AsyncTestCase):
             # Init mock client
             async with MockClient() as client:
                 # Mock out _post
-                client._post = MagicMock(return_value=return_fn())
+                client.post = MagicMock(return_value=return_fn())
 
                 expected_uri = "/supervised/predict"
 
                 # Make the call and assert exception raised
                 with self.assertRaises(Exception) as context:
-                    labels, probabilities = await client.predict(query, num_labels, threshold, headers=headers)
+                    labels, probabilities = await client.supervised.predict(query, num_labels, threshold, headers=headers)
                     self.assertIn("Invalid response for method 'get_sentence_vector'", str(context))
-                client._post.assert_called_with(expected_uri, data, headers=headers)
+                client.post.assert_called_with(expected_uri, data, headers=headers)
 
         self.run_async(async_test_function)
